+++ src/restfw/errors.py
@@ -41,7 +41,8 @@ def http_exception_to_dict(exc, request, include_status=False):
     }
     if include_status:
         result['status'] = status_code
-    if status_code == 404 and 'resource' not in detail:
+    if (status_code == 404 and exc.__class__.__name__ == 'HTTPNotFound'
+            and 'resource' not in detail):
         resource = getattr(request, 'context', None)
         if resource:
             elements = [request.view_name] if request.view_name else []